Allow getGallery to use different Isotope Gallieres
We have a Product Reader page where we have multiple galleries being displayed. We wanted the ability to pick a different Isotope Gallery (Store configuration > Galleries) so I've added to getGallery a second parameter which is null by default but can be the ID of which gallery you want. Below that, if the ID is detected it replaces the current on inside $arrConfig.

I've tested this and it works perfectly. Feel free to change to any style you'd like, but for a three line change this adds very helpful functionality.